Refactor ping_ok from datacenterlight.utils to utils.hosting_utils
