+++ datacenterlight/utils.py
@@ -1,5 +1,4 @@
 import logging
-import subprocess
 from django.contrib.sites.models import Site
 
 from datacenterlight.tasks import create_vm_task
@@ -95,21 +94,3 @@ def create_vm(billing_address_data, stripe_customer_id, specs,
                         'token', 'customer']:
         if session_var in request.session:
             del request.session[session_var]
-
-
-def ping_ok(host_ipv6):
-    """
-    A utility method to check if a host responds to ping requests. Note: the
-    function relies on `ping6` utility of debian to check.
-
-    :param host_ipv6 str type parameter that represets the ipv6 of the host to
-           checked
-    :return True if the host responds to ping else returns False
-    """
-    try:
-        output = subprocess.check_output("ping6 -c 1 -w 2 " + host_ipv6,
-                                         shell=True)
-    except Exception as ex:
-        logger.debug(host_ipv6 + " not reachable via ping. Error = " + str(ex))
-        return False
-    return True
